+++ src/Folder.php
@@ -248,6 +248,7 @@ class Folder {
      * @return array
      * @throws ConnectionFailedException
      * @throws Exceptions\InvalidMessageDateException
+     * @throws Exceptions\MessageNotFoundException
      * @throws Exceptions\RuntimeException
      */
     public function overview($sequence = null){
@@ -354,6 +355,7 @@ class Folder {
      * @throws Exceptions\RuntimeException
      * @throws Exceptions\EventNotFoundException
      * @throws Exceptions\MessageFlagException
+     * @throws Exceptions\MessageNotFoundException
      */
     public function idle(callable $callback, $timeout = 1200, $auto_reconnect = false) {
         $this->client->getConnection()->setConnectionTimeout($timeout);